main.swift에 있던 STEP 2 과제 따로 분리
