+++ CodeStarterCamp_Week1/main.swift
@@ -8,17 +8,12 @@
 
 import Foundation
 
-func makeIcecream(width: Int, height: Int, shape: String) {
-    for _ in 1...height {
-        print(String(repeating: shape, count: width))
-    }
-}
-
-func makeStick(height: Int, shape: String) {
-    for _ in 1...height {
-        print(String("    \(shape) \(shape)"))
-    }
-}
-
+/*STEP2
 makeIcecream(width: 11, height: 8, shape: "*")
-makeStick(height: 4, shape: "|")
+makeStick(height: 4, shape: "|")*/
+
+//STEP3
+printPepero(bodyLength: 10, body: "***", topping: " ", stickLength: 4)
+printPepero(bodyLength: 12, body: "***", topping: "&", stickLength: 4)
+printPepero(bodyLength: 12, body: "***", topping: "#", stickLength: 6)
+printPepero(bodyLength: 6, body: "|0|", topping: " ", stickLength: 4)
